feat(contract): consolidate JOB_STATES into @gamedevpl/contract
north-star-architecture.md Phase 1, fifth slice. apps/api/src/job-state.ts's
JOB_STATES (twelve internal job states) was hand-duplicated as apps/web's
BuildPhase, same identical values, same shape as SubmissionState/BuildStep
before it.

Moved JOB_STATES/JobState into packages/contract/src/job-state.ts. job-state.ts
keeps its state-machine logic (transitions, isTerminal, reconciliation) local
and re-exports the contract's values for its many existing importers.
submissionApi.ts drops the separate BuildPhase name entirely and uses
JobState directly -- nothing else in apps/web referenced BuildPhase by name,
only via `status.phase` field access, so the rename has no other call sites.
