+++ eslint-rules/module-size-baseline.json
@@ -874,7 +874,7 @@
     "packages/contract/src/agent-channel-routes.ts": 39,
     "packages/contract/src/gate-status.test.ts": 49,
     "packages/contract/src/gate-status.ts": 25,
-    "packages/contract/src/index.ts": 6,
+    "packages/contract/src/index.ts": 7,
     "packages/game-generator/src/engine/engine.test.ts": 35,
     "packages/game-generator/src/index.ts": 2,
     "packages/game-generator/src/mock.test.ts": 67,
