+++ packages/contract/src/index.ts
@@ -3,4 +3,5 @@
 export { AGENT_CHANNEL_ROUTES, type AgentChannelRouteKey, type AgentChannelRoutePath } from './agent-channel-routes.js';
 export { BUILD_EVENT_KINDS, BUILD_STEPS, type BuildEventKind, type BuildStep } from './build-event.js';
 export { deriveGateStatusString, derivePreviewGateStatus, GATE_STATUS_VALUES, type GateStatus } from './gate-status.js';
+export { JOB_STATES, type JobState } from './job-state.js';
 export { SUBMISSION_STATES, type SubmissionState } from './submission-state.js';
